move limits into seperate file
